Filesystem: Moved retarget related file interfaces into platform
