+++ platform/DirHandle.h
