+++ platform/FileBase.h
