+++ platform/FilePath.cpp
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-#include "drivers/FilePath.h"
+#include "platform/FilePath.h"
 
 namespace mbed {
