+++ platform/LocalFileSystem.cpp
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-#include "drivers/LocalFileSystem.h"
+#include "platform/LocalFileSystem.h"
 
 #if DEVICE_LOCALFILESYSTEM
